+++ dbparti/admin.py
@@ -1,30 +1,38 @@
 from django.contrib import admin
-from dbparti.utilities import DateTimeMixin
+from dbparti import backend
+from dbparti.backends.exceptions import PartitionColumnError, PartitionFilterError
 
 
-class PartitionableAdmin(DateTimeMixin, admin.ModelAdmin):
+class PartitionableAdmin(admin.ModelAdmin):
     partition_show = 'all'
 
     def __init__(self, *args, **kwargs):
-        """Initializes all the parent classes"""
-        admin.ModelAdmin.__init__(self, *args, **kwargs)
-        DateTimeMixin.__init__(
-            self,
-            partition_show=self.partition_show,
-            partition_range=self.opts.partition_range,
-            partition_column_type=self.opts.get_field(self.opts.partition_column).get_internal_type(),
-        )
+        super(PartitionableAdmin, self).__init__(*args, **kwargs)
+
+        if not self.opts.partition_column in self.opts.get_all_field_names():
+            raise PartitionColumnError(
+                model=self.opts.__dict__['object_name'],
+                current_value=self.opts.partition_column,
+                allowed_values=self.opts.get_all_field_names()
+            )
+
+        try:
+            self.filter = getattr(backend.filters, '{}PartitionFilter'.format(
+                self.opts.partition_type.capitalize()))(self.partition_show, **self.opts.__dict__)
+        except AttributeError:
+            import re
+            raise PartitionFilterError(
+                model=self.opts.__dict__['object_name'],
+                current_value=self.opts.partition_type,
+                allowed_values=[c.replace('PartitionFilter', '').lower() for c in dir(
+                    backend.filters) if re.match('\w+PartitionFilter', c) is not None and 'Base' not in c]
+            )
 
     def queryset(self, request):
         """Determines data from what partitions should be shown in django admin"""
-        fday, lday = self.get_partition_show_period(self.get_datetime_string('date'))
         qs = super(PartitionableAdmin, self).queryset(request)
 
-        if fday is None and lday is None:
-            qs = qs.all()
-        else:
-            qs = qs.filter(
-                **{self.opts.partition_column + '__gte': fday, self.opts.partition_column + '__lte': lday}
-            )
+        if self.partition_show != 'all':
+            qs = qs.extra(where=self.filter.apply())
 
         return qs